IB/{core,mlx5}: Support uhw definition per driver
Instead of supporting a generic unknown size of the uhw (driver
specific uhw_in and uhw_out) parameters via the common attributes,
declare the sizes specific to each driver in a driver specific tree.

When we initialize the parsing tree, this tree is merged with the
common parsing tree, creating this driver specific tree.
This tree could add driver specific types, actions and other
attributes as well.
We propose creating way of passing the merged parsing tree to the
user-space. By that, a user-space could query which features are
supported by the driver.

issue: 776663
Change-Id: I36125b694c1f51d23616f3080aeddfd70ca70998
Signed-off-by: Matan Barak <matanb@mellanox.com>