IB/core: Add macros for declaring actions and attributes
This patch adds macros for declaring action groups, actions,
attribute groups and attributes. These definitions are later
used by downstream patches to declare some of the common types.

In addition, we add some helper inline functions to copy_{from,to}
user-space buffers and check if an attribute is valid.

issue: 776663
Change-Id: I16a094e0da1b62e6a257d87a430040100da4fd7f
Signed-off-by: Matan Barak <matanb@mellanox.com>