IB/core: Add macros for declaring actions and attributes
This patch adds macros for declaring action groups, actions,
attribute groups and attributes. These definitions are later
used by downstream patches to declare some of the common types.

In addition, we add some helper inline functions to copy_{from,to}
user-space buffers and check if an attribute is valid.

issue: 776663
Change-Id: Ie16969d617f5907bd82237755e2ace1492bead21
Signed-off-by: Matan Barak <matanb@mellanox.com>