IB/core: Add support to finalize objects in one transaction
The new IOCTL based infrastructure either commits or rollbacks
all objects of the command as one transaction. In order to do
that, we introduce a notion of dealing with a collection of
objects that are related to a specific action.

This also requires adding a notion of an action and attribute.
An action contains a groups of attributes, where each group
contains several attributes.

When declaring these actions and attributes, we actually declare
their specifications. When a command is executed, we actually
allocates some space to hold auxiliary information.

issue: 776663
Change-Id: I5a96c0dc1cd97df994e7fa340ed6879055e1fd82
Signed-off-by: Matan Barak <matanb@mellanox.com>