IB/core: Add DEVICE type and root types structure
This adds the DEVICE type. This type support creating the context all
objects are created from. Moreover, it supports executing actions
which are related to the device itself, such as QUERY_DEVICE.
The only related object to this type is a singleton (per file
instance) context.

All standard types are put in the root structure.

issue: 776663
Change-Id: Idf6a7717957374a914b3fa90761a7e9743e5019b
Signed-off-by: Matan Barak <matanb@mellanox.com>