IB/core: Change completion channel to use the reworked objects schema
This patch adds the standard fd based type - completion_channel.
The completion_channel is now prefixed with ib_uobject, similarly
to the rest of the objects.
This required a few changes:
(1) We define a new completion channel fd based object type.
(2) completion_event and async_event are now two different types.
    The means they use different fops.
(3) We release the completion_channel exactly as we release other
    IDR based objects.
(4) Since ib_uobject is already kref-ed, we only add the kref to the
    async event.

A fd object requires filling out several parameters. Its op pointer
should point to uverbs_fd_ops and its size should be at least the
size if ib_uobject. We use a macro to make the type declaration
easier.

Change-Id: I415f4aa3309ea9543c0f16a1b9425470f19f7a72
Signed-off-by: Matan Barak <matanb@mellanox.com>