IB/core: Manage ib_ucontext with kref
Downstream patches in this series transform fd based objects (like
completion channel) to be standard user-objects. This means
user-objects could outlive the context. However, releasing objects
require a lock - as otherwise we could corrupt the objects' list.
This lock is part of the user-context. In order to mitigate this,
we add a kref onto the user-context.

Change-Id: I76a775faf100dd0bc8c9e7846e693223ea4f2658
Signed-off-by: Matan Barak <matanb@mellanox.com>