IB/core: Add legacy driver's user-data
In this phase, we don't want to change all the drivers to use
flexible driver's specific attributes. Therefore, we add two default
attributes: UHW_IN and UHW_OUT. These attributes are optional in some
commands and they encode the driver specific command data. We add
a function that extract this data and creates the legacy udata over
it.

issue: 776663
Change-Id: I21fad0b20fa47c328153ac71c7bfa8eb4d4c8063
Signed-off-by: Matan Barak <matanb@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>